Add configurable baudrate option
It is handy to be able to see the messages from rboot at a sensible
baudrate, so add an option to configure the uart divisor.

Notice that the APB is running at 52MHz in rboot instead of the normal 80MHz
in user applications.

Also notice that soft reboots (E.G. what you get by calling
system_restart()) does not reset the PLL or UART divisor, so detect this and
skip the UART setup so we don't end up with a wrong baudrate.